// src/lib/authz/spicedb-generated-types.ts
export enum ERP_DOMAIN_RELATIONS {
  MEMBER = 'member',
}

export enum ERP_DOMAIN_PERMISSIONS {
  IS_MEMBER = 'is_member',
}

export enum ERP_DOMAIN_SUBJECT_RELATIONS {
  USER_MEMBER = 'USER_MEMBER',
}

export const ERP_DOMAIN_SUBJECT_RELATIONS_MAP: SubjectRelationsMapType<
  RESOURCE_TYPES,
  ERP_DOMAIN_RELATIONS,
  ERP_DOMAIN_SUBJECT_RELATIONS
> = {
  USER_MEMBER: {
    relation: ERP_DOMAIN_RELATIONS.MEMBER,
    subjectType: RESOURCE_TYPES.ERP_USER,
  },
} as const;

export enum ERP_DOMAIN_SUBJECT_PERMISSIONS {
  USER_IS_MEMBER = 'USER_IS_MEMBER',
}

export const ERP_DOMAIN_SUBJECT_PERMISSIONS_MAP: SubjectPermissionsMapType<
  RESOURCE_TYPES,
  ERP_DOMAIN_PERMISSIONS,
  ERP_DOMAIN_SUBJECT_PERMISSIONS
> = {
  USER_IS_MEMBER: {
    permission: ERP_DOMAIN_PERMISSIONS.IS_MEMBER,
    subjectType: RESOURCE_TYPES.ERP_USER,
  },
};